Split out remoteLoginUser
